+++ 02-graphql-basic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# Dummy data nd queries
+
+- setup dummy data for users and posts
+- setup queries for users and posts
+- implement a `query` for getting back particular user or post
+
 # New types and resolvers, nodemon fix
 
 - add new types and resolvers